chore(change-test-directory): Modify test directory
