+++ client/__test__/__mocks__/mockLocalStorage.js
