+++ client/__test__/actions/AuthActions.test.js
