+++ client/__test__/component/user/NavBar.test.js
