+++ client/__test__/component/user/RegisterationForm.test.js
