+++ client/__test__/component/user/SideBar.test.js
