+++ client/__test__/reducers/AuthReducer.test.js
